+++ src/lines11.py
@@ -1,3 +1,5 @@
+# Triple-quoted strings, while loop
+
 REFRAIN = '''
 %d bottles of beer on the wall,
 %d bottles of beer,
