+++ src/lines13.py
@@ -1,14 +1,9 @@
-import doctest
+import unittest
 
+
+# Unit testing with unittest
 
 def median(pool):
-    '''Statistical median to demonstrate doctest.
-
-    >>> median([2, 9, 9, 7, 9, 2, 4, 5, 8])
-
-    6 #change to 7 in order to pass the test
-    '''
-
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
@@ -17,5 +12,10 @@ def median(pool):
         return (copy[int(size / 2 - 1)] + copy[int(size / 2)]) / 2
 
 
+class TestMedian(unittest.TestCase):
+    def testMedian(self):
+        self.assertEqual(median([2, 9, 9, 7, 9, 2, 4, 5, 8]), 7)
+
+
 if __name__ == '__main__':
-    doctest.testmod()
+    unittest.main()
