+++ src/lines18.py
@@ -1,3 +1,5 @@
+# 8-Queens Problem (recursion)
+
 BOARD_SIZE = 8
