+++ src/lines2.py
@@ -1,2 +1,4 @@
+# Input, assignment
+
 name = input('What is your name?\n')
 print('Hi, %s.' % name)
