+++ src/lines28.py
@@ -1,3 +1,5 @@
+# 8-Queens Problem (define your own exceptions)
+
 BOARD_SIZE = 8
