+++ src/lines3.py
@@ -1,3 +1,5 @@
+# For loop, built-in enumerate function, new style formatting
+
 friends = ['john', 'pat', 'gary', 'michael']
 for i, name in enumerate(friends):
     print("iteration {iteration} is {name}".format(iteration=i, name=name))
