+++ src/lines4.py
@@ -1,3 +1,5 @@
+# Fibonacci, tuple assignment
+
 parents, babies = (1, 1)
 while babies < 100:
     print('This generation has {0} babies'.format(babies))
