+++ src/lines5.py
@@ -1,3 +1,5 @@
+# Functions
+
 def greet(name):
     print('Hello, ', name)
