+++ src/lines7.py
@@ -1,3 +1,5 @@
+# Dictionaries, generator expressions
+
 prices = {'apple': 0.40, 'banana': 0.50}
 my_purchase = {
     'apple': 1,
