+++ src/lines8.py
@@ -1,5 +1,9 @@
 import sys
 
+# Command line arguments, exception handling
+# This program adds up integers that have
+# been passed as arguments in the command line
+
 try:
     total = sum(int(arg) for arg in sys.argv[1:])
     print('Sum = ', total)
