+++ src/lines9.py
@@ -1,5 +1,9 @@
 import glob
 
+# Opening files
+# indent your Python code to put into an email
+# glob supports Unix style pathname extensions
+
 python_files = glob.glob('*.py')
 for file_name in sorted(python_files):
     print('    -----' + file_name)
